+++ console/controllers/RunController.php
@@ -16,6 +16,7 @@ use common\models\HelpModel;
 use common\models\InternalClass;
 use common\models\MoscowPoint;
 use common\models\Motorcycle;
+use common\models\NewsSubscription;
 use common\models\Notice;
 use common\models\Participant;
 use common\models\Region;
@@ -24,6 +25,7 @@ use common\models\RequestForSpecialStage;
 use common\models\SpecialChamp;
 use common\models\SpecialStage;
 use common\models\Stage;
+use common\models\SubscriptionQueue;
 use common\models\Time;
 use common\models\TmpAthlete;
 use common\models\TmpFigureResult;
@@ -771,236 +773,6 @@ class RunController extends Controller
 		return true;
 	}
 	
-	public function actionChangeChampionshipsStatus()
-	{
-		$time = time();
-		$count = 0;
-		
-		//чемпионат прошел
-		$championships = Championship::findAll(['status' => Championship::STATUS_PRESENT]);
-		foreach ($championships as $championship) {
-			$stages = $championship->stages;
-			/** @var Stage $stage */
-			foreach ($stages as $stage) {
-				if (!$stage->dateOfThe || $stage->dateOfThe > $time) {
-					continue 2;
-				}
-			}
-			$championship->status = Championship::STATUS_PAST;
-			$championship->save();
-			$count++;
-		}
-		
-		$championships = SpecialChamp::findAll(['status' => SpecialChamp::STATUS_PRESENT]);
-		foreach ($championships as $championship) {
-			$stages = $championship->stages;
-			/** @var SpecialStage $stage */
-			foreach ($stages as $stage) {
-				if (!$stage->dateEnd || $stage->dateEnd > $time) {
-					continue 2;
-				}
-			}
-			$championship->status = SpecialChamp::STATUS_PAST;
-			$championship->save();
-			$count++;
-		}
-		
-		//чемпионат начался
-		$championships = Championship::findAll(['status' => Championship::STATUS_UPCOMING]);
-		foreach ($championships as $championship) {
-			if (Stage::find()->where(['championshipId' => $championship->id])
-				->andWhere(['<=', 'dateOfThe', $time])->one()
-			) {
-				$championship->status = Championship::STATUS_PRESENT;
-				$championship->save();
-				$count++;
-			}
-		}
-		
-		$championships = SpecialChamp::findAll(['status' => SpecialChamp::STATUS_UPCOMING]);
-		foreach ($championships as $championship) {
-			if (SpecialStage::find()->where(['championshipId' => $championship->id])
-				->andWhere(['<=', 'dateEnd', $time])->one()
-			) {
-				$championship->status = SpecialChamp::STATUS_PRESENT;
-				$championship->save();
-				$count++;
-			}
-		}
-		
-		//чемпионат завершился и снова начался
-		$year = Year::findOne(['year' => date('Y')]);
-		if ($year) {
-			$championships = Championship::findAll(['status' => Championship::STATUS_PAST, 'yearId' => $year->id]);
-			foreach ($championships as $championship) {
-				if (Stage::find()->where(['championshipId' => $championship->id])
-					->andWhere(['>=', 'dateOfThe', $time])->one()
-				) {
-					$championship->status = Championship::STATUS_PRESENT;
-					$championship->save();
-					$count++;
-				}
-			}
-			
-			$championships = SpecialChamp::findAll(['status' => SpecialChamp::STATUS_PAST, 'yearId' => $year->id]);
-			foreach ($championships as $championship) {
-				if (SpecialStage::find()->where(['championshipId' => $championship->id])
-					->andWhere(['>=', 'dateEnd', $time])->one()
-				) {
-					$championship->status = Championship::STATUS_PRESENT;
-					$championship->save();
-					$count++;
-				}
-			}
-		}
-		
-		echo 'Change ' . $count . ' items';
-		
-		return true;
-	}
-	
-	public function actionChangeStagesStatus()
-	{
-		$time = time();
-		//открыта регистрация на этап
-		Stage::updateAll(['status' => Stage::STATUS_START_REGISTRATION], [
-			'and',
-			['status' => Stage::STATUS_UPCOMING],
-			['not', ['startRegistration' => null]],
-			['<=', 'startRegistration', $time]
-		]);
-		
-		//завершена регистрация на этап
-		Stage::updateAll(['status' => Stage::STATUS_END_REGISTRATION], [
-			'and',
-			['status' => Stage::STATUS_START_REGISTRATION],
-			['not', ['endRegistration' => null]],
-			['<=', 'endRegistration', $time]
-		]);
-		
-		//текущий этап
-		Stage::updateAll(['status' => Stage::STATUS_PRESENT], [
-			'and',
-			['<=', 'dateOfThe', $time],
-			['not', ['dateOfThe' => null]],
-			['status' => [Stage::STATUS_UPCOMING, Stage::STATUS_END_REGISTRATION, Stage::STATUS_END_REGISTRATION]]
-		]);
-		
-		//прошедший этап
-		/** @var Stage[] $stages */
-		$stages = Stage::find()->where(['not', ['status' => Stage::STATUS_PAST]])->all();
-		foreach ($stages as $stage) {
-			if ($stage->dateOfThe && ($stage->dateOfThe + 86400) <= $time) {
-				$stage->status = Stage::STATUS_PAST;
-				$stage->save();
-			}
-		}
-		
-		//Приём результатов
-		SpecialStage::updateAll(['status' => SpecialStage::STATUS_START], [
-			'and',
-			['status' => SpecialStage::STATUS_UPCOMING],
-			['not', ['dateStart' => null]],
-			['<=', 'dateStart', $time]
-		]);
-		
-		//Приём результатов завершен
-		SpecialStage::updateAll(['status' => SpecialStage::STATUS_CALCULATE_RESULTS], [
-			'and',
-			['status' => SpecialStage::STATUS_START],
-			['not', ['dateEnd' => null]],
-			['<=', 'dateEnd', $time]
-		]);
-		
-		//прошедший этап
-		/** @var SpecialStage[] $stages */
-		$stages = SpecialStage::find()->where(['not', ['status' => SpecialStage::STATUS_PAST]])->all();
-		foreach ($stages as $stage) {
-			if ($stage->dateEnd && ($stage->dateEnd + 86400) <= $time) {
-				$stage->status = Stage::STATUS_PAST;
-				$stage->save();
-			}
-		}
-		
-		return true;
-	}
-	
-	public function actionChangePhotoStatus()
-	{
-		Stage::updateAll(['trackPhotoStatus' => Stage::PHOTO_PUBLISH], [
-			'and',
-			['not', ['trackPhoto' => null]],
-			['trackPhotoStatus' => Stage::PHOTO_NOT_PUBLISH],
-			['status' => Stage::STATUS_PAST]
-		]);
-		
-		return true;
-	}
-	
-	public function actionCheckSize()
-	{
-		exec('df -h', $output, $return_var);
-		if ($output) {
-			if (!isset($output[1])) {
-				$errors = new Error();
-				$errors->text = 'Невозможно проверить остаток дискового пространства на хостинге';
-				$errors->save();
-				
-				if (YII_ENV == 'prod') {
-					$text = 'Невозможно проверить остаток дискового пространства на хостинге';
-					\Yii::$app->mailer->compose('text', ['text' => $text])
-						->setTo('nadia__@bk.ru')
-						->setFrom(['support@gymkhana-cup.ru' => 'GymkhanaCup'])
-						->setSubject('gymkhana-cup.ru: ошибка на сайте')
-						->send();
-				}
-				
-				return false;
-			}
-			$string = $output[1];
-			$array = explode('G', $string);
-			if (!isset($array[2])) {
-				$errors = new Error();
-				$errors->text = 'Невозможно проверить остаток дискового пространства на хостинге';
-				$errors->save();
-				
-				if (YII_ENV == 'prod') {
-					$text = 'Невозможно проверить остаток дискового пространства на хостинге';
-					\Yii::$app->mailer->compose('text', ['text' => $text])
-						->setTo('nadia__@bk.ru')
-						->setFrom(['support@gymkhana-cup.ru' => 'GymkhanaCup'])
-						->setSubject('gymkhana-cup.ru: ошибка на сайте')
-						->send();
-				}
-				
-				return false;
-			}
-			$size = trim($array[2]);
-			echo $size . PHP_EOL;
-			if ($size < 1) {
-				$errors = new Error();
-				$errors->type = Error::TYPE_CRITICAL_ERROR;
-				$errors->text = 'На хостинге осталось менее 1GB свободного места';
-				$errors->save();
-			} elseif ($size <= 2) {
-				$errors = new Error();
-				$errors->text = 'На хостинге осталось ' . $size . 'GB свободного места';
-				$errors->type = Error::TYPE_SIZE;
-				$errors->save();
-				
-				if (YII_ENV == 'prod') {
-					\Yii::$app->mailer->compose('text', ['text' => $errors->text])
-						->setTo('nadia__@bk.ru')
-						->setFrom(['support@gymkhana-cup.ru' => 'GymkhanaCup'])
-						->setSubject('gymkhana-cup.ru: ошибка на сайте')
-						->send();
-				}
-			}
-		}
-		
-		return true;
-	}
-	
 	public function actionMailTest()
 	{
 		\Yii::$app->mailer->compose('text', ['text' => 'проверка почты'])
@@ -1008,28 +780,6 @@ class RunController extends Controller
 			->setFrom(['support@gymkhana-cup.ru' => 'GymkhanaCup'])
 			->setSubject('gymkhana-cup.ru: проверка почты')
 			->send();
-	}
-	
-	public function actionSendNotice($athleteId)
-	{
-		$text = 'У вас необычный случай с результатом GP 8 – ваш рейтинг находится на границе классов. В связи с этим, вопрос о вашем переводе в новый класс вынесен на рассмотрение. Как только решение будет принято, вам придёт уведомление о присвоении нового класса.';
-		var_dump(Notice::add($athleteId, $text));
-		
-		return true;
-	}
-	
-	public function actionAddRecordTimes()
-	{
-		$results = FigureTime::find()->where(['>', 'percent', 100])->all();
-		/** @var FigureTime $result */
-		foreach ($results as $result) {
-			$figure = $result->figure;
-			$result->recordInMoment = $figure->bestTime;
-			$result->save(false);
-		}
-		echo 'Update ' . count($results) . ' records' . PHP_EOL;
-		
-		return true;
 	}
 	
 	public function actionIndices()
@@ -1132,45 +882,6 @@ class RunController extends Controller
 		} catch (\Throwable $ex) {
 			var_dump($ex->getMessage());
 		}
-	}
-	
-	public function actionAddVideoLinks()
-	{
-		/** @var TmpFigureResult[] $tmp */
-		$tmp = TmpFigureResult::find()->where(['isNew' => 0])->andWhere(['not', ['videoLink' => null]])->all();
-		$count = 0;
-		foreach ($tmp as $tmpItem) {
-			$needAdd = false;
-			if (mb_strstr($tmpItem->videoLink, 'http://', 'UTF-8') !== false
-				|| mb_strstr($tmpItem->videoLink, 'https://', 'UTF-8') !== false
-			) {
-				if (mb_strstr($tmpItem->videoLink, 'http://vk.', 'UTF-8') !== false
-					|| mb_strstr($tmpItem->videoLink, 'https://vk.', 'UTF-8') !== false
-				) {
-					if (mb_strstr($tmpItem->videoLink, 'video', 'UTF-8') !== false) {
-						$needAdd = true;
-					}
-				} else {
-					$needAdd = true;
-				}
-			}
-			if ($needAdd) {
-				$figureResult = FigureTime::findOne(['id' => $tmpItem->figureResultId]);
-				if ($figureResult) {
-					$figureResult->videoLink = $tmpItem->videoLink;
-					if (!$figureResult->save()) {
-						var_dump($figureResult->errors);
-						
-						return false;
-					}
-					$count++;
-				}
-			}
-		}
-		
-		echo 'update ' . $count . ' items';
-		
-		return true;
 	}
 	
 	public function actionInsertMoscowPoints()
@@ -1384,17 +1095,6 @@ class RunController extends Controller
 		return true;
 	}
 	
-	public function actionTestTime()
-	{
-		$timeHuman = '00:00.191';
-		$time = HelpModel::convertTime($timeHuman);
-		echo 'Old: ' . $time . ', ' . HelpModel::convertTimeToHuman($time) . PHP_EOL;
-		$referenceTime = round($time / 10) * 10;
-		echo 'New: ' . $referenceTime . ', ' . HelpModel::convertTimeToHuman($referenceTime) . PHP_EOL;
-		
-		return true;
-	}
-	
 	public function actionFixForStages($stageId)
 	{
 		$stage = Stage::findOne($stageId);
@@ -1414,52 +1114,6 @@ class RunController extends Controller
 						$participant->percent . ', new ' . $percent . PHP_EOL, FILE_APPEND);
 				}
 			}
-		}
-		
-		return true;
-	}
-	
-	public function actionUpdateEraser()
-	{
-		$figure = Figure::findOne(['title' => 'Eraser']);
-		$times = FigureTime::findAll(['figureId' => $figure->id]);
-		foreach ($times as $time) {
-			$percent = round($time->resultTime / $figure->bestTime * 100, 2);
-			$time->actualPercent = $percent;
-			$time->needClassCalculate = false;
-			if (!$time->recordInMoment) {
-				$time->recordInMoment = $time->time;
-			}
-			$time->save(false);
-		}
-		if (!$figure->severalRecords) {
-			$figure->severalRecords = 1;
-			$figure->save(false);
-		}
-	}
-	
-	public function actionAddQualifications()
-	{
-		FigureTime::updateAll(['stageId' => null]);
-		/** @var Stage[] $stages */
-		$stages = Stage::find()->all();
-		foreach ($stages as $stage) {
-			$count = 0;
-			$dateStart = $stage->dateOfThe;
-			$dateEnd = $stage->dateOfThe + 12 * 3600;
-			/** @var FigureTime[] $figureTimes */
-			$figureTimes = FigureTime::find()
-				->where(['>=', 'date', $dateStart])->andWhere(['<=', 'date', $dateEnd])->andWhere(['stageId' => null])->all();
-			foreach ($figureTimes as $item) {
-				if (Participant::find()->where(['athleteId' => $item->athleteId, 'motorcycleId' => $item->motorcycleId,
-				                                'stageId'   => $stage->id])->one()
-				) {
-					$item->stageId = $stage->id;
-					$item->save(false);
-					$count++;
-				}
-			}
-			echo $stage->id . '. ' . $stage->title . ' - ' . $count . PHP_EOL;
 		}
 		
 		return true;
